backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-binance==1.0.19
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
python-telegram-bot==20.6
discord-webhook==1.3.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
websockets==12.0
ccxt==4.4.1
pandas==2.1.3
numpy==1.26.2
requests>=2.31.0
slowapi==0.1.8
bcrypt==4.1.2
cryptography==41.0.0
psutil